requirements.txt
django
djangorestframework
pillow
psycopg2
